// src/api/model/listWorkspaceResponse.ts
/**
 * Onepanel
 * Onepanel API
 *
 * The version of the OpenAPI document: 0.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Workspace } from './workspace';


export interface ListWorkspaceResponse { 
    count?: number;
    workspaces?: Array<Workspace>;
    page?: number;
    pages?: number;
    totalCount?: number;
}